pyproject.toml
[project]
name = "chinese-text-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "loguru>=0.7.3",
    "tqdm>=4.66.0",
]
